+++ main.rb
@@ -1,5 +1,5 @@
 require './app'
-
+require './optional.rb'
 def list_options
   'Please choose an option by entering a number:
     1 - List of all books
@@ -11,36 +11,19 @@ def list_options
     7 - Exit'
 end
 
-def option(option, app) # rubocop:disable Metrics/CyclomaticComplexity
-  case option
-  when 1
-    app.list_books
-  when 2
-    app.list_people
-  when 3
-    app.create_person
-  when 4
-    app.create_book
-  when 5
-    app.create_rental
-  when 6
-    app.list_rentals
-  when 7
-    exit
-  else
-    puts 'Invalid option, please type correct number!'
-  end
+def intro
+  "Welcome to School Library App!\n\n"
 end
 
 def main
   app = App.new
 
-  puts "Welcome to School Library App!\n\n"
+  puts intro
 
   loop do
     puts list_options
     option = gets.chomp.to_i
-    option(option, app)
+    Option.option(option, app)
   end
 end
